models modified

+++ src/app/models/JobPostSkillSet-model.ts
@@ -1,6 +1,9 @@
+import { JobPost } from './job-post-model'
+import { SkillSet } from './SkillSet-model'
+
 export class JobPostSkillSet {
     id: BigInteger
     jobPost: JobPost
-    skillLevel: int
+    skillLevel: BigInt
     skillSet: SkillSet
 }
